pyproject.toml
[project]
name = "nexagent"
version = "0.1.0"
description = "Autonomous trading agent powered by Nexwave signals"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Nexwave", email = "hello@nexwave.so" }]
dependencies = [
    "ccxt>=4.3",
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "typer>=0.12",
    "pydantic-settings>=2.2",
    "aiosqlite>=0.20",
    "httpx>=0.27",
    "rich>=13",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
alerts = ["python-telegram-bot>=21"]
x402 = ["solders>=0.21", "base58>=2"]
dev = ["pytest>=8", "pytest-asyncio>=0.23", "httpx"]

[project.scripts]
nex = "nexagent.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
